backend/requirements.txt
django
django-environ
django-cors-headers
djangorestframework
pillow
django-storages
django-ckeditor
# psycopg2
psycopg2-binary
whitenoise
dj_database_url
coreapi
transformers
torch